fix(termcap): update g_num_capabilities type to size_t for consistency

+++ include/internal/ft_termcap_internal.h
@@ -6,7 +6,7 @@
 /*   By: rel-qoqu <rel-qoqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/17 13:29:56 by rel-qoqu          #+#    #+#             */
-/*   Updated: 2025/09/17 14:04:51 by rel-qoqu         ###   ########.fr       */
+/*   Updated: 2025/09/17 16:27:28 by rel-qoqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_ft_termcap_entry {
 }	t_ft_termcap_entry;
 
 extern t_ft_termcap_entry	*g_capabilities;
-extern int32_t				g_num_capabilities;
+extern size_t				g_num_capabilities;
 
 bool	parse_termcap_line(const char *line);
